Move `StoreManager` to production
